fix procdor postprocess throwing exceptions (CERT-Polska#355)

+++ drakcore/drakcore/util.py
@@ -1,5 +1,8 @@
 import base64
 import os
+import sys
+
+from karton2 import Config
 
 
 def find_config():
@@ -14,21 +17,43 @@ def find_config():
         raise RuntimeError("Configuration file was not found neither in {} nor {}".format(local_path, etc_path))
 
 
+def get_config():
+    cfg = Config(find_config())
+
+    try:
+        access_key = cfg.config['minio']['access_key']
+        secret_key = cfg.config['minio']['secret_key']
+    except KeyError:
+        sys.stderr.write('WARNING! Misconfiguration: section [minio] of config.ini doesn\'t contain access_key or secret_key.\n')
+        return cfg
+
+    if not access_key and not secret_key:
+        if not os.path.exists('/etc/drakcore/minio.env'):
+            raise RuntimeError('ERROR! MinIO access credentials are not configured (and can not be auto-detected), unable to start.\n')
+
+        with open('/etc/drakcore/minio.env', 'r') as f:
+            minio_cfg = [line.strip().split('=', 1) for line in f if line.strip() and '=' in line]
+            minio_cfg = {k: v for k, v in minio_cfg}
+
+        try:
+            cfg.config['minio']['access_key'] = minio_cfg['MINIO_ACCESS_KEY']
+            cfg.config['minio']['secret_key'] = minio_cfg['MINIO_SECRET_KEY']
+            cfg.minio_config = dict(cfg.config.items("minio"))
+        except KeyError:
+            sys.stderr.write('WARNING! Misconfiguration: minio.env doesn\'t contain MINIO_ACCESS_KEY or MINIO_SECRET_KEY.\n')
+
+    return cfg
+
+
 def setup_config():
-    print('Generating MinIO access key and secret key...')
+    if os.path.exists('/etc/drakcore/minio.env'):
+        print('MinIO environment file already exists, skipping...')
+        return
+
+    print('Generating MinIO environment file...')
     access_key = base64.b64encode(os.urandom(30)).decode('ascii').replace('+', '-').replace('/', '_')
     secret_key = base64.b64encode(os.urandom(30)).decode('ascii').replace('+', '-').replace('/', '_')
 
-    with open('/etc/drakcore/config.ini', 'r') as f:
-        data = f.read()
-        data = data.replace('{MINIO_ACCESS_KEY}', access_key).replace('{MINIO_SECRET_KEY}', secret_key)
-
-    with open('/etc/drakcore/config.ini', 'w') as f:
-        f.write(data)
-
-    with open('/etc/drakcore/minio.env', 'r') as f:
-        data = f.read()
-        data = data.replace('{MINIO_ACCESS_KEY}', access_key).replace('{MINIO_SECRET_KEY}', secret_key)
-
     with open('/etc/drakcore/minio.env', 'w') as f:
-        f.write(data)
+        f.write(f'MINIO_ACCESS_KEY={access_key}\n')
+        f.write(f'MINIO_SECRET_KEY={secret_key}\n')
